services/requirements.txt
psycopg2-binary>=2.9.9
google-api-python-client==2.88.0
google-auth==2.17.3
requests==2.31.0
